+++ src/main.ts
@@ -1,6 +1,8 @@
 export { useRoutes, useMatch, usePathParams } from './router'
 export type { RouteParams, RouteOptionParams, PathParamOptions } from './router'
 
+export { RouterProvider } from './context'
+
 export { useRedirect, Redirect } from './redirect'
 export type { RedirectProps, UseRedirectProps } from './redirect'
